requirements.txt
discord.py==2.5.2
opencv_python==4.12.0.88
Pillow==11.3.0
psutil==7.0.0
pyautogui==0.9.54
pycryptodomex==3.23.0
pyttsx3==2.99
pywin32==311
Requests==2.32.4
